refactor: share the canvas plugin test fixtures
content and scene were byte-identical in three specs, and the
seed-normalize-select preamble appeared four times.
